backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.4
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.2.1
httpx==0.26.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
